Ranges (pegjs/pegjs#30): Implement ranges support. Range syntax: ``` expression|  exact | expression|   ..   | expression|min..   | expression|   ..max| expression|min..max| ```
Introduce two new opcodes:
* IF_LT <min>, <then part length>, <else part length>
* IF_GE <max>, <then part length>, <else part length>

Introduce a new AST node -- `repeated`, that contains expression and the minimum and maximum number of it repetition.
If `node.min.value` is `null` or isn't positive -- check of the minimum length isn't made.
If `node.max.value` is `null`, check of the maximum length isn't made.
If `node.min` is `null` then it is equals to the `node.max` (exact repetitions case)