Optimization (inference match result): eliminate unnecessary checks for result of parse, when it can be statically determined
The idea in that we calculate expected rules result and use that information when generate bytecode.
Some constructs of the PEG grammar never fail, for example, optional operator `?` or repetition
operator `*`. So it is unnecessary to check their result for the failure.

New pass `inference-match-result.js` adds to the each AST node new integer property `match`, which
represents possible result of matching:
- `<0` - node is never match, for example, `!('a'*)` (negation of the always matched node)
- `=0` - sometimes node matched, sometimes not. This is behavior before that change
- `>0` - node is always matched, for example, `'a'*` (because result is or empty array, or array with some elements)