Cleanup bytecode generator (pegjs/pegjs#450): Move function code generating to `generateJs` pass from `generateBytecode` pass.
That allows to generate parser on other languages and more easy replace actions.
Also, this is helpful for better hiding of the peggy internals from an initializers
and functions code, which is needed for the import feature.

After this commit, a function code (semantic predicates and actions) is stored
in the `functions` field of the grammar node instead of the `consts` field.

This is also opens doors for supporting source maps, because we get a place for
the storing information about the original positions of the codes