Add a pass to simplify single-character choices
I have a modified version of examples/xml.peggy in my project, and
was trying to optimize it for speed. I found a few places where
it could be rewritten to avoid backtracking, getting about a 20%
win, but with some readability cost.

Then I noticed that there are a number of long lists of alternate
character classes, and they end up being implemented very
inefficiently; effectively it tries one regex after another in
sequence.

I manually combined all the character classes for BaseChar into
a single character class, and got a nearly 2x speed up.

So I've added a pass that does that automatically. The result is
a more than 3x speedup parsing the xml files that I work with
(which seem fairly typical).